backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.0
httpx>=0.27.0
anthropic>=0.28.0
pydantic>=2.7.0
pydantic-settings>=2.3.0
sqlalchemy>=2.0.30
asyncpg>=0.29.0
alembic>=1.13.1
redis>=5.0.4
PyJWT>=2.8.0
passlib[bcrypt]>=1.7.4
email-validator>=2.1.0
